requirements.txt
streamlit
ultralytics
pillow
torch
requests
groq
python-dotenv
